delay mailbox modifications by 4 seconds to make undo work w/o hitting server
our query overwrites can only delete items; so if we delete and then undo it takes a full round trip + sync for the item to reappear.

To avoid this we delay all calls that insert query overwrites by 4 seconds so a subsequent undo call will just replace the previous job.

4 seconds has been choosen because it is slighly longer than the long snackbar timeout.