provide undo snackbar for moveToTrash
Most other operations implement by undo by sipmly executing the opposite Mua
action. (For example archive->moveToInbox)

However there is no oppososite action to moveToTrash as a thread could be in
muliple mailboxes. Therefor it is better to just attempt to cancel the Job
(which is started with a 4 second delay).

This commit also fixes an issues with QueryItemOverwrites where toggleing
a keyword and then toggling it again and the removing the corresponding
thread from a query while being offline causes the thread to briefly
reappear.

Example: Mark as read, Mark as unread, archive. The second action would
cancel out the first action; thus removing the overwrites before the third
action could be executed.

(Previously we attempted to fix this by just marking an overwrite as executed
instead of deleting it; however that isn’t good either because it leaves
the overwrites in the db until the next query refresh)

This commit 'properly' fixes this by introducing different kind of query item
overwrites (one for keywords, one for mailboxes, and in the future one for
important)