create query for anyNotInRole()
This is a remainder from a failed attempt at keeping emails in important
when moving them to trash.
The general idea was that any email that is in Trash is considered as being
in Trash regardless of what other mailboxes they are still in.
However that breaks JMAP in multiple ways; for example the EmailQuery
inMailboxOtherThan can not be used with this model. More importantly
the mailbox counts don’t work any more.

closes #2
closes #3