add splash screen
initializing the app is rather expensive (looking at you Loggers); therefore we
display a splash screen in front of the LttrsActivity.

We also use the Main activity to query whether or not we have accounts on the
main thread and use that information to either redirect into Setup or Lttrs.
While doing i/o on the main (UI) thread is generally a bad idea we ignore this
rule since we prefer that to a double redirect (To Lttrs to Setup).
The exists query should be fairly inexpensive anyway. (Also the overall
wait-time for the user wouldn’t be lower with a double redirect (we just
wouldn’t block))