use email from query result to display timestamp
We use server side ordering of query results. Threads are ordered by the time
stamp of the first matching email (not by the most recent email in that thread)

So a thread where only the first email is in a certain mailbox will be ordered
by the timestamp of that email. Therefor when displaying the timestamp we need
to pick that email.