refactor email modification worker behaviour
This commit changes the behaviour of workers that modify email.

Previously we set unique names for workers that take the thread id into
account. This was in an effort to ensure that setKeyword(seen, true) and
setKeyword(seen, false) cancel each other out. Additionally we set some
time delay (that was higher than a snackbar timeout) in an effort for undone
actions to never hit the server. (Instead the jobs would just cancel each other)

However uniqueness by thread-id would make jobs run in parallel which would
cause some nasty failed attempts due to state missmatch errors - we tried
to circumvent that by forcing WorkManager to use a single thread model.

This was all very complicated.

Now Workers are unique by account-id (meaning all jobs per account run after
each other) and run instantly.

This downside of this is that you see a flicker if delayed jobs first set and
then unset a keyword. The first job will delete the overwrite. And then a users
sees the keyword turned on and then turned off again.

(This happens if the users makes the change and the undo in airplane mode) and
then keeps looking at the screen for the changes.

However we deliberatly accept this in order to have code that is a lot cleaner.

(No single thread, no delay, etc)